+++ sw_stack/discover/sles15/src/2024.04.00/build_2_check.sh
@@ -1,0 +1,11 @@
+#!/bin/bash
+
+source ./basics.sh
+
+echo $DSLSW_INSTALL_DIR
+echo `which $FC`
+echo `which $CC`
+
+echo $LD_LIBRARY_PATH
+
+rm -rf build