+++ sw_stack/discover/sles15/src/2024.04.00/build_3_inode-clean.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+source ./basics.sh
+
+rm -rf $DSLSW_BASE